backend/requirements.txt
Django==2.1
psycopg2
psycopg2-binary==2.7.5
django-extensions==2.1.0
djangorestframework==3.8.2
djangorestframework-simplejwt==4.2.0
djoser==1.5.1
django-cors-headers==2.4.0
django-crontab==0.7.1
slackclient==1.3.1
Sphinx==2.0.1